onet2r: Work with 'O*NET' Web Services, Archives, and Weights

Access the 'O*NET' Web Services Application Programming Interface (API) version 2.0. Search occupations, retrieve occupation details such as skills, knowledge, abilities, tasks, and work activities, access database tables, perform taxonomy crosswalks, join wage and employment context, assemble longitudinal archive panels, and validate user-supplied occupational measures. Results are returned as tibbles for easy integration with tidyverse workflows.

See also

Author

Maintainer: Alex Farach alexfarach@microsoft.com [copyright holder]