Get All O\*NET Occupation Skills (details)

Retrieves all skill rows for a single occupation by automatically paginating through the endpoint.

Usage

onet_skills_all(code, page_size = 2000, show_progress = TRUE)

Arguments

code

An O\*NET-SOC occupation code.

page_size

Integer specifying how many rows to fetch per request (default 2000, which is the API maximum).

show_progress

Logical indicating whether to show progress messages for pagination (default TRUE).

Value

A tibble of skills.

Examples

onet_skills_all("15-1252.00", show_progress = FALSE)
#> # A tibble: 35 × 5
#>    id      related                                  name  description importance
#>    <chr>   <chr>                                    <chr> <chr>            <int>
#>  1 2.B.3.e https://api-v2.onetcenter.org/online/oc… Prog… Writing co…         75
#>  2 2.A.2.a https://api-v2.onetcenter.org/online/oc… Crit… Using logi…         72
#>  3 2.B.4.e https://api-v2.onetcenter.org/online/oc… Judg… Considerin…         66
#>  4 2.A.2.b https://api-v2.onetcenter.org/online/oc… Acti… Understand…         63
#>  5 2.A.1.a https://api-v2.onetcenter.org/online/oc… Read… Understand…         63
#>  6 2.B.4.g https://api-v2.onetcenter.org/online/oc… Syst… Determinin…         63
#>  7 2.A.1.b https://api-v2.onetcenter.org/online/oc… Acti… Giving ful…         60
#>  8 2.B.2.i https://api-v2.onetcenter.org/online/oc… Comp… Identifyin…         60
#>  9 2.B.4.h https://api-v2.onetcenter.org/online/oc… Syst… Identifyin…         60
#> 10 2.B.3.b https://api-v2.onetcenter.org/online/oc… Tech… Generating…         60
#> 11 2.B.3.a https://api-v2.onetcenter.org/online/oc… Oper… Analyzing …         56
#> 12 2.A.1.c https://api-v2.onetcenter.org/online/oc… Writ… Communicat…         56
#> 13 2.A.1.d https://api-v2.onetcenter.org/online/oc… Spea… Talking to…         53
#> 14 2.B.5.a https://api-v2.onetcenter.org/online/oc… Time… Managing o…         53
#> 15 2.B.1.b https://api-v2.onetcenter.org/online/oc… Coor… Adjusting …         50
#> 16 2.A.2.d https://api-v2.onetcenter.org/online/oc… Moni… Monitoring…         50
#> 17 2.B.1.e https://api-v2.onetcenter.org/online/oc… Inst… Teaching o…         47
#> 18 2.B.1.f https://api-v2.onetcenter.org/online/oc… Serv… Actively l…         47
#> 19 2.B.1.a https://api-v2.onetcenter.org/online/oc… Soci… Being awar…         47
#> 20 2.A.1.e https://api-v2.onetcenter.org/online/oc… Math… Using math…         44
#> 21 2.B.3.g https://api-v2.onetcenter.org/online/oc… Oper… Watching g…         44
#> 22 2.B.1.c https://api-v2.onetcenter.org/online/oc… Pers… Persuading…         44
#> 23 2.B.3.m https://api-v2.onetcenter.org/online/oc… Qual… Conducting…         44
#> 24 2.A.2.c https://api-v2.onetcenter.org/online/oc… Lear… Selecting …         41
#> 25 2.B.1.d https://api-v2.onetcenter.org/online/oc… Nego… Bringing o…         41
#> 26 2.B.3.k https://api-v2.onetcenter.org/online/oc… Trou… Determinin…         41
#> 27 2.B.5.d https://api-v2.onetcenter.org/online/oc… Mana… Motivating…         31
#> 28 2.B.3.c https://api-v2.onetcenter.org/online/oc… Equi… Determinin…         28
#> 29 2.A.1.f https://api-v2.onetcenter.org/online/oc… Scie… Using scie…         28
#> 30 2.B.3.h https://api-v2.onetcenter.org/online/oc… Oper… Controllin…         25
#> 31 2.B.3.d https://api-v2.onetcenter.org/online/oc… Inst… Installing…         22
#> 32 2.B.5.b https://api-v2.onetcenter.org/online/oc… Mana… Determinin…         19
#> 33 2.B.5.c https://api-v2.onetcenter.org/online/oc… Mana… Obtaining …         16
#> 34 2.B.3.j https://api-v2.onetcenter.org/online/oc… Equi… Performing…          6
#> 35 2.B.3.l https://api-v2.onetcenter.org/online/oc… Repa… Repairing …          3