Get All O*NET Occupation Work Activities (details)

Retrieves all work activity rows for a single occupation by automatically paginating through the endpoint.

Usage

onet_work_activities_all(code, page_size = 2000, show_progress = TRUE)

Arguments

code

An O*NET-SOC occupation code.

page_size

Integer specifying how many rows to fetch per request (default 2000, which is the API maximum).

show_progress

Logical indicating whether to show progress messages for pagination (default TRUE).

Value

A tibble of work activities.

Examples

onet_work_activities_all("15-1252.00", show_progress = FALSE)
#> # A tibble: 41 × 5
#>    id        related                                name  description importance
#>    <chr>     <chr>                                  <chr> <chr>            <int>
#>  1 4.A.3.b.1 https://api-v2.onetcenter.org/online/… Work… Using comp…         90
#>  2 4.A.2.a.2 https://api-v2.onetcenter.org/online/… Proc… Compiling,…         85
#>  3 4.A.2.b.1 https://api-v2.onetcenter.org/online/… Maki… Analyzing …         84
#>  4 4.A.2.b.2 https://api-v2.onetcenter.org/online/… Thin… Developing…         83
#>  5 4.A.4.a.2 https://api-v2.onetcenter.org/online/… Comm… Providing …         81
#>  6 4.A.2.a.4 https://api-v2.onetcenter.org/online/… Anal… Identifyin…         80
#>  7 4.A.2.b.3 https://api-v2.onetcenter.org/online/… Upda… Keeping up…         78
#>  8 4.A.2.b.6 https://api-v2.onetcenter.org/online/… Orga… Developing…         77
#>  9 4.A.1.a.1 https://api-v2.onetcenter.org/online/… Gett… Observing,…         76
#> 10 4.A.2.a.3 https://api-v2.onetcenter.org/online/… Eval… Using rele…         74
#> 11 4.A.3.b.6 https://api-v2.onetcenter.org/online/… Docu… Entering, …         70
#> 12 4.A.2.b.4 https://api-v2.onetcenter.org/online/… Deve… Establishi…         69
#> 13 4.A.1.a.2 https://api-v2.onetcenter.org/online/… Moni… Monitoring…         68
#> 14 4.A.1.b.1 https://api-v2.onetcenter.org/online/… Iden… Identifyin…         67
#> 15 4.A.4.a.1 https://api-v2.onetcenter.org/online/… Inte… Translatin…         65
#> 16 4.A.4.a.4 https://api-v2.onetcenter.org/online/… Esta… Developing…         64
#> 17 4.A.2.b.5 https://api-v2.onetcenter.org/online/… Sche… Scheduling…         58
#> 18 4.A.4.b.6 https://api-v2.onetcenter.org/online/… Prov… Providing …         57
#> 19 4.A.4.b.3 https://api-v2.onetcenter.org/online/… Trai… Identifyin…         46
#> 20 4.A.4.b.5 https://api-v2.onetcenter.org/online/… Coac… Identifyin…         43
#> 21 4.A.1.b.3 https://api-v2.onetcenter.org/online/… Esti… Estimating…         42
#> 22 4.A.2.a.1 https://api-v2.onetcenter.org/online/… Judg… Assessing …         40
#> 23 4.A.4.a.3 https://api-v2.onetcenter.org/online/… Comm… Communicat…         38
#> 24 4.A.4.a.7 https://api-v2.onetcenter.org/online/… Reso… Handling c…         35
#> 25 4.A.4.b.1 https://api-v2.onetcenter.org/online/… Coor… Getting me…         34
#> 26 4.A.4.c.1 https://api-v2.onetcenter.org/online/… Perf… Performing…         33
#> 27 4.A.4.a.5 https://api-v2.onetcenter.org/online/… Assi… Providing …         32
#> 28 4.A.3.a.3 https://api-v2.onetcenter.org/online/… Cont… Using eith…         32
#> 29 4.A.4.b.2 https://api-v2.onetcenter.org/online/… Deve… Encouragin…         31
#> 30 4.A.4.c.3 https://api-v2.onetcenter.org/online/… Moni… Monitoring…         28
#> 31 4.A.1.b.2 https://api-v2.onetcenter.org/online/… Insp… Inspecting…         27
#> 32 4.A.3.b.2 https://api-v2.onetcenter.org/online/… Draf… Providing …         17
#> 33 4.A.4.b.4 https://api-v2.onetcenter.org/online/… Guid… Providing …         17
#> 34 4.A.3.a.1 https://api-v2.onetcenter.org/online/… Perf… Performing…         17
#> 35 4.A.3.b.5 https://api-v2.onetcenter.org/online/… Repa… Servicing,…         16
#> 36 4.A.4.a.6 https://api-v2.onetcenter.org/online/… Sell… Convincing…         15
#> 37 4.A.4.c.2 https://api-v2.onetcenter.org/online/… Staf… Recruiting…         13
#> 38 4.A.3.a.2 https://api-v2.onetcenter.org/online/… Hand… Using hand…          5
#> 39 4.A.3.a.4 https://api-v2.onetcenter.org/online/… Oper… Running, m…          3
#> 40 4.A.4.a.8 https://api-v2.onetcenter.org/online/… Perf… Performing…          3
#> 41 4.A.3.b.4 https://api-v2.onetcenter.org/online/… Repa… Servicing,…          2